requirements.txt
Django>=4.2
psycopg2-binary
Pillow
django-userforeignkey
xhtml2pdf
djangorestframework